% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_splicing_variants_and_annotation.R
\name{create_splicing_variants_and_annotation}
\alias{create_splicing_variants_and_annotation}
\title{Internal function to create alternative splicing events}
\usage{
create_splicing_variants_and_annotation(
  gtf_path,
  valid_chromosomes,
  event_probs,
  outdir,
  ncores,
  write_gff,
  max_genes
)
}
\arguments{
\item{gtf_path}{path to gtf from which exon_supersets are created}

\item{ncores}{the number of cores to be utilized for parallel generation
of splice variants.}

\item{max_genes}{}
}
\value{
the exons, junctions and retained introns as data table in gtf style
}
\description{
This is not intended to be called directly;
instead it is meant to be called via \code{\link{simulate_alternative_cplicing}}
}
