% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_preset.R
\docType{data}
\name{presets}
\alias{presets}
\title{Parameter presets for the \code{\link{simulate_alternative_splicing}} function}
\format{
An object of class \code{list} of length 3.
}
\usage{
data(presets)
}
\description{
Parameter presets for the \code{\link{simulate_alternative_splicing}} function
}
\author{
Quirin Manz \email{quirin.manz@tum.de}
}
\keyword{data}
